requirements.txt
agate==1.7.1
annotated-types==0.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
blinker==1.7.0
boto3==1.34.92
botocore==1.34.92
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
confluent-kafka==2.3.0
crashtest==0.4.1
cryptography==42.0.5
dash==2.16.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbt-core==1.7.13
dbt-extractor==0.5.1
dbt-postgres==1.7.13
dbt-semantic-interfaces==0.4.4
debugpy==1.8.1
decorator==5.1.1
diskcache==5.6.3
distlib==0.3.8
dulwich==0.21.7
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib-metadata==6.11.0
installer==0.7.0
ipykernel==6.29.4
ipython==8.23.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kafka==1.3.5
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
keyring==24.3.1
leather==0.4.0
llama_cpp_python==0.2.64
Logbook==1.5.3
MarkupSafe==2.1.5
mashumaro==3.12
matplotlib-inline==0.1.7
minimal-snowplow-tracker==0.0.2
more-itertools==10.2.0
msgpack==1.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parsedatetime==2.6
parso==0.8.4
pathspec==0.11.2
pexpect==4.9.0
pip-review==1.3.0
pkginfo==1.10.0
platformdirs==4.2.1
plotly==5.21.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyOpenSSL==24.1.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
rapidfuzz==3.8.1
referencing==0.35.0
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
rpds-py==0.18.0
s3transfer==0.10.1
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.5.0
stack-data==0.6.3
tenacity==8.2.3
text-unidecode==1.3
tomlkit==0.12.4
tornado==6.4
traitlets==5.14.3
trove-classifiers==2024.4.10
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
virtualenv==20.26.0
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.43.0
xattr==1.1.0
zipp==3.18.1
